Unit test script now cleans up any messy files. Added endpoint for searching by trend (recent, hot, popular) and by page. Refactored some endpoint unit tests.

+++ run_tests.py
@@ -1,13 +1,25 @@
 import os
 import unittest
 
+prev = None
+if 'FLASK_DATABASE_URI' in os.environ:
+    prev = os.environ['FLASK_DATABASE_URI']
+
 os.environ['FLASK_DATABASE_URI'] = "project/db/:memory:"
 
-loader = unittest.TestLoader()
-start_dir = 'project/tests'
-suite = loader.discover(start_dir)
+def clean():
+    if os.path.isfile(os.path.abspath(os.environ['FLASK_DATABASE_URI'])):
+        os.remove(os.path.abspath(os.environ['FLASK_DATABASE_URI']))
+    if os.path.isfile(os.path.abspath(os.environ['FLASK_DATABASE_URI'] + '-journal')):
+        os.remove(os.path.abspath(os.environ['FLASK_DATABASE_URI'] + '-journal'))
+try:
+    loader = unittest.TestLoader()
+    start_dir = 'project/tests'
+    suite = loader.discover(start_dir)
 
-runner = unittest.TextTestRunner()
-runner.run(suite)
-
-os.remove(os.path.abspath(os.environ['FLASK_DATABASE_URI']))
+    runner = unittest.TextTestRunner()
+    runner.run(suite)
+finally: #clean up db files whether errored or not
+    clean()
+    if prev:
+        os.environ['FLASK_DATABASE_URI'] = prev
